+++ tools/xdhq/examples/common/index.php
@@ -1,0 +1,9 @@
+<?php
+// Is used by 'php -S localhost:8080'.
+// Is only used by the production release of the Atlas toolkit
+// NOT used during development (the 'Node.js' web server is used instead).
+
+require( "h:\\hg\\epeios\\tools\\xdhwebq\\ZND\\XDHWebQ.php");
+
+echo XDHWebQ::serve();
+?>