+++ tools/xdhwebq/NJS/XDHWebQ.js
@@ -115,6 +115,7 @@ function prolog() {
 		'<!DOCTYPE html>',
 		'<html>',
 		'	<head>',
+		'		<meta name="generator" content="XDHWebQNJS"/>',
 		'		<meta charset="UTF-8" />',
 		'		<meta http-equiv="X-UA-Compatible" content="IE=edge" />',
 		userHead(),